fix: screenshot bug (when screen is close to border)
